==> 01_datafilterST.ipynb <==
"""
First notebook to open, explore, and filter full ST dataset from Cantin

Shaina Lu
Zador & Gillis Labs
March 2020
"""

%matplotlib inline
import matplotlib.pyplot as plt
import h5py
import pandas as pd
import numpy as np
import seaborn as sns

# Read in Full ST Data

EXPRS_FILE = "/data/slu/cantin_ST/exprmat.tsv"
SPOTSMETA_FILE = "/data/slu/cantin_ST/spotstable.tsv"
EXPTMETA_FILE = "/data/slu/cantin_ST/slicestable.tsv"
CLUSTER_FILE = "/data/slu/cantin_ST/cluster-id-name.csv"
RES_FILE = "/data/slu/cantin_ST/res-29.tsv"

exprsdata = pd.read_csv(EXPRS_FILE, sep='\t',header=0, index_col=0, quoting=3)

exprsdata.head()

exprsdata.shape

spotsmeta = pd.read_csv(SPOTSMETA_FILE, sep='\t',header=0, index_col=0, quoting=3)
exptmeta = pd.read_csv(EXPTMETA_FILE, sep='\t',header=0, index_col=0, quoting=3)

print(spotsmeta.shape)
spotsmeta.head()

print(exptmeta.shape)
exptmeta.head()

#look that the remaining two files
clusteridname = pd.read_csv(CLUSTER_FILE, sep=';',header=0, index_col=0, quoting=3)
res29 = pd.read_csv(RES_FILE, sep='\t',header=None, index_col=0, quoting=3)

print(clusteridname.shape)
clusteridname.head()

print(res29.shape)
res29.head()

# propagate ontology

__read in ontology data__

ONTOLOGY_PATH = "/data/slu/allen_adult_mouse_ISH/ontologyABA.csv"

#read in ontology file
ontology = pd.read_csv(ONTOLOGY_PATH)
ontology = ontology.drop([ontology.columns[5], ontology.columns[6]], axis=1)
ontology = ontology.fillna(-1)  #make root's parent -1

ontology.head()

__first, propagate raw ontology__

#do this in a sloppy way that i know works
#propogate first then look-up
#goes up to the point where all labels are root
proponto = pd.DataFrame()
proponto[['id','1']] = ontology[['id', 'parent']]

curr = 1
while True:
    #get series representing next level up
    newcol = []
    for i in range(proponto.shape[0]):
        if proponto.iloc[i,curr] == -1:   #if already at parent of root, just give root again (-1)
            newval = -1
        else:
            newval = ontology.loc[ontology["id"] == int(proponto.iloc[i,curr]), "parent"].item()
        newcol.append(newval)

    if len(proponto[str(curr)].value_counts()) < 3:  #all are root after this point
        break
    proponto[str(curr+1)] = pd.Series(newcol)     #add series to df
    curr += 1

#at 10 levels up, all labels are root
#proponto.head()
proponto.loc[:,'10'].value_counts()

proponto = proponto.astype(int)  #typecast as integer
proponto = proponto.replace(to_replace=-1, value=997)  #replace root's parent (-1) with root (997)
proponto.head()

__now propagate full binary ontology for the ST spots themselves__

#initiate full binary ontology dataframe
fullontST = pd.DataFrame(index=exprsdata.index, columns=ontology.id, dtype=np.int64)
fullontST.columns = fullontST.columns.astype(str)
fullontST = fullontST.fillna(0) #make all entries 0

for i in range(len(spotsmeta.acronym.unique())): #for each unique brain area
    currareaid = int(ontology.loc[ontology.acronym == spotsmeta.acronym.unique()[i], "id"].item())
    tempareas = proponto.loc[proponto.id == currareaid, :] #get row representing that area propagated

    for val in tempareas.values.flatten():
        #set row,col entries to 1 
        fullontST.loc[spotsmeta.acronym == spotsmeta.acronym.unique()[i], str(val)] = 1

fullontST.head()
fullontST.shape

__remove spots that map to ventricular systems of fiber tracts__

#ventricular systesm - 251 spots map here
print(fullontST.loc[:,"73"].sum())
#fiber tracts - 3072 map here
print(fullontST.loc[:,"1009"].sum())

#get the ventricular and fibertract spots
vsspots = fullontST.loc[fullontST["73"]==1, "73"].index
fibspots = fullontST.loc[fullontST["1009"]==1, "1009"].index

#remove the ventricular and fiber tract spots
fullontST.drop(labels=vsspots, axis=0, inplace=True)
exprsdata.drop(labels=vsspots, axis=0, inplace=True)
spotsmeta.drop(labels=vsspots, axis=0, inplace=True)

fullontST.drop(labels=fibspots, axis=0, inplace=True)
exprsdata.drop(labels=fibspots, axis=0, inplace=True)
spotsmeta.drop(labels=fibspots, axis=0, inplace=True)

print(fullontST.shape)
print(exprsdata.shape)
print(spotsmeta.shape)

# Filter

__remove spots that do not map to the brain__ 

#spotsmeta.acronym.unique()

It appears that all spots here do map to the brain 

__remove genes that only contain expression in less than X% of samples__

#count number of spots for each gene that has a non-zero exprs value
nonzerocount = exprsdata.astype(bool).sum(axis=0)

#filter for genes having expression in at least 0.1% of samples
exprsdata_filt = exprsdata.loc[:, nonzerocount > (0.001*exprsdata.shape[0])]

exprsdata_filt.shape

# add id col to meta data

lots of my code from before uses the brain id column, so instead of re-doing this 

temp = dict(zip(ontology.loc[:,"acronym"], ontology.loc[:,"id"]))
ids = []
for val in spotsmeta.acronym.values:
    ids.append(temp[val])

spotsmeta["id"] = ids

spotsmeta.head()
spotsmeta.shape

# Write to new hdf5

only write the exprsdata and spotsmeta dataframes since the other three files are unchanged from cantin \
also write propagated ontology

OUT_PATH = "cantin_ST_filt_v2.h5"

exprsdata_filt.to_hdf(OUT_PATH, key="STspots", mode='w')
spotsmeta.to_hdf(OUT_PATH, key="STspotsmeta")
fullontST.to_hdf(OUT_PATH, key="STpropont")
