==> 14_CFScombinefiles.ipynb <==
"""
combine all output files from 13_CFScrossdataset.py that was split up when run to increase comp time

Shaina Lu
Zador & Gillis Labs
May 2020
"""

import pandas as pd
import numpy as np

# ST to ABA

def read4files(file1, file2, file3, file4):
    file1 = pd.read_csv(file1, sep=',', index_col=False)
    file1.set_index(file1.columns, inplace=True)
    
    file2 = pd.read_csv(file2, sep=',', index_col=False)
    file2.set_index(file2.columns, inplace=True)
    
    file3 = pd.read_csv(file3, sep=',', index_col=False)
    file3.set_index(file3.columns, inplace=True)
    
    file4 = pd.read_csv(file4, sep=',', index_col=False)
    file4.set_index(file4.columns, inplace=True)
    
    return file1, file2, file3, file4

def combine4df(df1, df2, df3, df4):
    df1.iloc[101:201,:] = df2.iloc[101:201,:]
    df1.iloc[201:301,:] = df3.iloc[201:301,:]
    df1.iloc[301:len(list(df1)),:] = df4.iloc[301:len(list(df1)),:]
    
    return df1

ST train

file1 = "/home/slu/spatial/data/cross_CFS/STtoABA_STtrain_f1_CFS_052220.csv"
file2 = "/home/slu/spatial/data/cross_CFS/STtoABA_STtrain_f1_2_CFS_052420.csv"
file3 = "/home/slu/spatial/data/cross_CFS/STtoABA_STtrain_f1_3_CFS_052420.csv"
file4 = "/home/slu/spatial/data/cross_CFS/STtoABA_STtrain_f1_4_CFS_052420.csv"
outfile = "/home/slu/spatial/data/cross_CFS/STtoABA_STtrain_f1_all_CFS_052920.csv"

data1, data2, data3, data4 = read4files(file1, file2, file3, file4)
allcombined = combine4df(data1, data2, data3, data4)

allcombined.shape

allcombined.iloc[99:105]

allcombined.to_csv(outfile, header=True, index=False)

ST test

file1 = "/home/slu/spatial/data/cross_CFS/STtoABA_STtest_f1_CFS_052220.csv"
file2 = "/home/slu/spatial/data/cross_CFS/STtoABA_STtest_f1_2_CFS_052420.csv"
file3 = "/home/slu/spatial/data/cross_CFS/STtoABA_STtest_f1_3_CFS_052420.csv"
file4 = "/home/slu/spatial/data/cross_CFS/STtoABA_STtest_f1_4_CFS_052420.csv"
outfile = "/home/slu/spatial/data/cross_CFS/STtoABA_STtest_f1_all_CFS_052920.csv"

data1, data2, data3, data4 = read4files(file1, file2, file3, file4)
allcombined = combine4df(data1, data2, data3, data4)

allcombined.shape

allcombined.to_csv(outfile, header=True, index=False)

featsets ST train

file1 = "/home/slu/spatial/data/cross_CFS/STtoABA_featsetsSTtrain_f1_CFS_052220.csv"
file2 = "/home/slu/spatial/data/cross_CFS/STtoABA_featsetsSTtrain_f1_2_CFS_052420.csv"
file3 = "/home/slu/spatial/data/cross_CFS/STtoABA_featsetsSTtrain_f1_3_CFS_052420.csv"
file4 = "/home/slu/spatial/data/cross_CFS/STtoABA_featsetsSTtrain_f1_4_CFS_052420.csv"
outfile = "/home/slu/spatial/data/cross_CFS/STtoABA_featsetsSTtrain_f1_all_CFS_052920.csv"

data1, data2, data3, data4 = read4files(file1, file2, file3, file4)
allcombined = combine4df(data1, data2, data3, data4)

allcombined.shape

allcombined.to_csv(outfile, header=True, index=False)

ABA all

file1 = "/home/slu/spatial/data/cross_CFS/STtoABA_ABAall_f1_CFS_052220.csv"
file2 = "/home/slu/spatial/data/cross_CFS/STtoABA_ABAall_f1_2_CFS_052420.csv"
file3 = "/home/slu/spatial/data/cross_CFS/STtoABA_ABAall_f1_3_CFS_052420.csv"
file4 = "/home/slu/spatial/data/cross_CFS/STtoABA_ABAall_f1_4_CFS_052420.csv"
outfile = "/home/slu/spatial/data/cross_CFS/STtoABA_ABAall_f1_all_CFS_052920.csv"

data1, data2, data3, data4 = read4files(file1, file2, file3, file4)
allcombined = combine4df(data1, data2, data3, data4)

allcombined.shape

allcombined.to_csv(outfile, header=True, index=False)

# ABA to ST

def read5files(file1, file2, file3, file4, file5):
    file1 = pd.read_csv(file1, sep=',', index_col=False)
    file1.set_index(file1.columns, inplace=True)
    
    file2 = pd.read_csv(file2, sep=',', index_col=False)
    file2.set_index(file2.columns, inplace=True)
    
    file3 = pd.read_csv(file3, sep=',', index_col=False)
    file3.set_index(file3.columns, inplace=True)
    
    file4 = pd.read_csv(file4, sep=',', index_col=False)
    file4.set_index(file4.columns, inplace=True)
    
    file5 = pd.read_csv(file5, sep=',', index_col=False)
    file5.set_index(file5.columns, inplace=True)
    
    return file1, file2, file3, file4, file5

def combine5df(df1, df2, df3, df4, df5):
    df1.iloc[91:181,:] = df2.iloc[91:181,:]
    df1.iloc[181:271,:] = df3.iloc[181:271,:]
    df1.iloc[271:361,:] = df4.iloc[271:361,:]
    df1.iloc[361:len(list(df1)),:] = df5.iloc[361:len(list(df1)),:]
    
    return df1

ABA train

file1 = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtrain_f1_1_CFS_052420.csv"
file2 = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtrain_f1_2_CFS_052420.csv"
file3 = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtrain_f1_3_CFS_052420.csv"
file4 = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtrain_f1_4_CFS_052420.csv"
file5 = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtrain_f1_5_CFS_052420.csv"
outfile = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtrain_f1_all_CFS_052920.csv"

data1, data2, data3, data4, data5 = read5files(file1, file2, file3, file4, file5)
allcombined = combine5df(data1, data2, data3, data4, data5)

allcombined.to_csv(outfile, header=True, index=False)

ABA test

file1 = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtest_f1_1_CFS_052420.csv"
file2 = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtest_f1_2_CFS_052420.csv"
file3 = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtest_f1_3_CFS_052420.csv"
file4 = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtest_f1_4_CFS_052420.csv"
file5 = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtest_f1_5_CFS_052420.csv"
outfile = "/home/slu/spatial/data/cross_CFS/ABAtoST_ABAtest_f1_all_CFS_052920.csv"

data1, data2, data3, data4, data5 = read5files(file1, file2, file3, file4, file5)
allcombined = combine5df(data1, data2, data3, data4, data5)

allcombined.to_csv(outfile, header=True, index=False)

featsets ABA train

file1 = "/home/slu/spatial/data/cross_CFS/ABAtoST_featsetsABAtrain_f1_1_CFS_052420.csv"
file2 = "/home/slu/spatial/data/cross_CFS/ABAtoST_featsetsABAtrain_f1_2_CFS_052420.csv"
file3 = "/home/slu/spatial/data/cross_CFS/ABAtoST_featsetsABAtrain_f1_3_CFS_052420.csv"
file4 = "/home/slu/spatial/data/cross_CFS/ABAtoST_featsetsABAtrain_f1_4_CFS_052420.csv"
file5 = "/home/slu/spatial/data/cross_CFS/ABAtoST_featsetsABAtrain_f1_5_CFS_052420.csv"
outfile = "/home/slu/spatial/data/cross_CFS/ABAtoST_featstsABAtrain_f1_all_CFS_052920.csv"

data1, data2, data3, data4, data5 = read5files(file1, file2, file3, file4, file5)
allcombined = combine5df(data1, data2, data3, data4, data5)

allcombined.to_csv(outfile, header=True, index=False)

ST all

file1 = "/home/slu/spatial/data/cross_CFS/ABAtoST_STall_f1_1_CFS_052420.csv"
file2 = "/home/slu/spatial/data/cross_CFS/ABAtoST_STall_f1_2_CFS_052420.csv"
file3 = "/home/slu/spatial/data/cross_CFS/ABAtoST_STall_f1_3_CFS_052420.csv"
file4 = "/home/slu/spatial/data/cross_CFS/ABAtoST_STall_f1_4_CFS_052420.csv"
file5 = "/home/slu/spatial/data/cross_CFS/ABAtoST_STall_f1_5_CFS_052420.csv"
outfile = "/home/slu/spatial/data/cross_CFS/ABAtoST_STall_f1_all_CFS_052920.csv"

data1, data2, data3, data4, data5 = read5files(file1, file2, file3, file4, file5)
allcombined = combine5df(data1, data2, data3, data4, data5)

allcombined.to_csv(outfile, header=True, index=False)
